+++ packages/contracts/deploy/deploy.ts
@@ -8,7 +8,7 @@ const TEST_ROUTERS = [
   "0x627306090abaB3A6e1400e9345bC60c78a8BEf57", // local router
 ];
 
-const SKIP_SETUP = [1, 56, 250, 137, 100];
+const SKIP_SETUP = [1, 56, 250, 137, 100, 42161];
 
 /**
  * Hardhat task defining the contract deployments for nxtp