+++ packages/contracts/hardhat.config.ts
@@ -15,6 +15,7 @@ import "./src/tasks/mintTestToken";
 import "./src/tasks/setupTestRouter";
 import "./src/tasks/getChainId";
 import "./src/tasks/renounceOwnership";
+import "./src/tasks/proposeTransferOwnership";
 
 dotEnvConfig();
 
@@ -122,6 +123,11 @@ const config: HardhatUserConfig = {
       accounts: { mnemonic },
       chainId: 1287,
       url: "https://moonbeam-alpha.api.onfinality.io/public",
+    },
+    "arbitrum-one": {
+      accounts: { mnemonic },
+      chainId: 42161,
+      url: "https://arb1.arbitrum.io/rpc",
     },
     fuji: {
       accounts: { mnemonic },